refactor: split graphql.rs (2400 lines) into module directory
Pure mechanical split — no logic changes:
- mod.rs: re-exports, build_schema, KoanSchema, DbHandle
- types.rs: all Gql* structs/enums with Object/SimpleObject impls
- queries.rs: QueryRoot with all query resolvers
- mutations.rs: MutationRoot with all mutation resolvers
- helpers.rs: paginate, year/genre/favourite helpers, downloads
- server.rs: cmd_serve, cmd_serve_daemon, execute_in_process

Removed unused Edge import and pub use re-export.
